requirements.txt
kafka-python==2.1.5
python-dotenv==1.1.0
websockets==15.0.1
